requirements.txt
fastapi
uvicorn
motor
transformers
datasets
python-dotenv